+++ Graph/Graph/src/com/gmail/at/ivanehreshi/graph/PathFinder.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 public interface PathFinder {
 	public static int INF = Integer.MAX_VALUE - 1000;
 	boolean findAll();
-	boolean relax(int u, int v, int w);
+	boolean relax(int u, int v, double w);
 	public ArrayList<Integer> getPath(int to);
-	public int[] getDist();
+	public double[] getDist();
 }
